+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-- Fix Fabric crash.
+- The `Colorize Own Name` option now applies to the player list.
+- Added `Show Head Borders` option, which renders player heads on the locator bar without the colored border if disabled.
 
 **Full Changelog**: https://github.com/BrandonItaly/Locator-Colors/commits/master